pyproject.toml
[tool.poetry]
name = "bt-sec-analyzer"
version = "0.1.0"
description = "Enterprise-grade Bluetooth security testing framework for authorized assessments"
authors = ["mfscpayload-690"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/mfscpayload-690/bt-sec-analyzer"
repository = "https://github.com/mfscpayload-690/bt-sec-analyzer"
keywords = ["bluetooth", "security", "testing", "pentest", "wireless"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Information Technology",
    "Topic :: Security",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
]

[tool.poetry.dependencies]
python = "^3.12"
pybluez = "^0.30"
bleak = "^0.22.0"
structlog = "^24.1.0"
ollama = "^0.1.0"
pyyaml = "^6.0"
click = "^8.1.7"
rich = "^13.7.0"
reportlab = "^4.0.9"
weasyprint = "^60.2"
aiofiles = "^23.2.1"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
psutil = "^5.9.8"
cryptography = "^42.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.3"
pytest-mock = "^3.12.0"
black = "^24.1.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
isort = "^5.13.2"
pylint = "^3.0.3"
pre-commit = "^3.6.0"

[tool.poetry.scripts]
bt-sec-analyzer = "bt_sectester.__main__:main"
bt-sec-analyzer-cli = "bt_sectester.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "requires_bluetooth: marks tests that require Bluetooth hardware",
    "requires_root: marks tests that require root/sudo privileges",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
strict_optional = true

[tool.coverage.run]
source = ["bt_sectester"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
